Abstract config files
