+++ gulpfile.js/config/browserSync.js
@@ -1,0 +1,8 @@
+var config = require('./')
+
+module.exports = {
+  server: {
+    baseDir: config.publicDirectory
+  },
+  files: ['pubilc/**/*.html']
+}
