+++ gulpfile.js/config/deploy.js
@@ -1,0 +1,6 @@
+var config = require('./')
+
+module.exports = {
+  url: 'http://greypants.github.io/gulp-starter/',
+  src: config.publicDirectory + '/**/*'
+}
