+++ gulpfile.js/config/iconFont.js
@@ -1,0 +1,17 @@
+var config = require('./')
+
+module.exports = {
+  name: 'Gulp Starter Icons',
+  src: config.sourceAssets + '/icons/*.svg',
+  dest: config.publicAssets + '/fonts',
+  sassDest: config.sourceAssets + '/stylesheets/generated',
+  template: './gulpfile.js/tasks/iconFont/template.sass',
+  sassOutputName: '_icons.sass',
+  fontPath: '../fonts',
+  className: 'icon',
+  options: {
+    fontName: 'icons',
+    appendCodepoints: true,
+    normalize: false
+  }
+}
