+++ gulpfile.js/config/images.js
@@ -1,0 +1,6 @@
+var config = require('./')
+
+module.exports = {
+  src: config.sourceAssets + "/images/**",
+  dest: config.publicAssets + "/images"
+}
