+++ gulpfile.js/config/index.js
@@ -1,0 +1,8 @@
+var config = {}
+
+config.publicDirectory = "./public";
+config.sourceDirectory = "./app";
+config.publicAssets    = config.publicDirectory + "/assets";
+config.sourceAssets    = config.sourceDirectory + "/assets";
+
+module.exports = config;
