+++ gulpfile.js/config/sass.js
@@ -1,0 +1,11 @@
+var config = require('./')
+
+module.exports = {
+  autoprefixer: { browsers: ['last 2 version'] },
+  src: config.sourceAssets + "/stylesheets/**/*.{sass,scss}",
+  dest: config.publicAssets + '/stylesheets',
+  settings: {
+    indentedSyntax: true, // Enable .sass syntax!
+    imagePath: 'assets/images' // Used by the image-url helper
+  }
+}
