+++ gulpfile.js/tasks/browserSync.js
@@ -1,7 +1,7 @@
 var browserSync = require('browser-sync');
 var gulp        = require('gulp');
-var config      = require('../config').browserSync;
+var config      = require('../config/browserSync')
 
-gulp.task('browserSync', ['webpack:development'], function() {
-  browserSync(config);
+gulp.task('browserSync', function() {
+  return browserSync(config);
 });
