--- gulpfile.js/tasks/build.js
@@ -1,6 +1,0 @@
-var gulp         = require('gulp');
-var gulpSequence = require('gulp-sequence');
-
-gulp.task('build', function(cb) {
-  gulpSequence('clean', ['iconFont', 'images'], ['sass', 'webpack:production'], 'html', cb);
-});
