+++ gulpfile.js/tasks/default.js
@@ -1,3 +1,3 @@
 var gulp = require('gulp');
 
-gulp.task('default', ['build', 'watch']);
+gulp.task('default', ['build:development']);
