+++ gulpfile.js/tasks/iconFont/generateIconSass.js
@@ -1,5 +1,5 @@
 var gulp         = require('gulp');
-var config       = require('../../config').iconFont;
+var config       = require('../../config/iconFont');
 var swig         = require('gulp-swig');
 var rename       = require('gulp-rename');
 var handleErrors = require('../../lib/handleErrors');
