+++ gulpfile.js/tasks/iconFont/index.js
@@ -1,6 +1,6 @@
 var gulp             = require('gulp');
 var iconfont         = require('gulp-iconfont');
-var config           = require('../../config').iconFont;
+var config           = require('../../config/iconFont');
 var generateIconSass = require('./generateIconSass');
 var handleErrors     = require('../../lib/handleErrors');
