+++ gulpfile.js/tasks/images.js
@@ -1,6 +1,6 @@
 var browserSync = require('browser-sync');
 var changed     = require('gulp-changed');
-var config      = require('../config').images;
+var config      = require('../config/images');
 var gulp        = require('gulp');
 var imagemin    = require('gulp-imagemin');
