--- gulpfile.js/tasks/production-build.js
@@ -1,6 +1,0 @@
-var gulp         = require('gulp');
-var gulpSequence = require('gulp-sequence');
-
-gulp.task('production-build', function(cb) {
-  gulpSequence('karma', 'build', 'rev', cb);
-});
