+++ gulpfile.js/tasks/sass.js
@@ -3,7 +3,7 @@ var browserSync  = require('browser-sync');
 var sass         = require('gulp-sass');
 var sourcemaps   = require('gulp-sourcemaps');
 var handleErrors = require('../lib/handleErrors');
-var config       = require('../config').sass;
+var config       = require('../config/sass');
 var autoprefixer = require('gulp-autoprefixer');
 
 gulp.task('sass', function () {
