+++ gulpfile.js/tasks/watch.js
@@ -1,10 +1,13 @@
-var gulp   = require('gulp');
-var config = require('../config');
-var watch  = require('gulp-watch');
+var gulp     = require('gulp');
+var html     = require('../config/html');
+var iconFont = require('../config/iconFont');
+var images   = require('../config/images');
+var sass     = require('../config/sass');
+var watch    = require('gulp-watch');
 
-gulp.task('watch', ['build', 'webpack:development', 'browserSync'], function(callback) {
-  watch(config.images.src, function() { gulp.start('images'); });
-  watch(config.sass.src, function() { gulp.start('sass'); });
-  watch(config.iconFont.src, function() { gulp.start('iconFont'); });
-  watch(config.html.watch, function() { gulp.start('html'); });
+gulp.task('watch', ['browserSync'], function() {
+  watch(images.src, function() { gulp.start('images'); });
+  watch(sass.src, function() { gulp.start('sass'); });
+  watch(iconFont.src, function() { gulp.start('iconFont'); });
+  watch(html.watch, function() { gulp.start('html'); });
 });
