+++ karma.conf.js
@@ -1,5 +1,5 @@
-var karmaConfig = require('./gulpfile.js/config').karma;
+var karmaConfig = require('./gulpfile.js/config/karma')
 
 module.exports = function(config) {
-  config.set(karmaConfig);
-};
+  config.set(karmaConfig)
+}
