+++ package.json
@@ -11,7 +11,7 @@
     "url": "git://github.com/greypants/gulp-starter.git"
   },
   "scripts": {
-    "postinstall": "gulp build",
+    "postinstall": "gulp",
     "test": "karma start"
   },
   "dependencies": {
@@ -48,7 +48,7 @@
     "karma-nyan-reporter": "0.0.50",
     "karma-sinon-chai": "^0.3.0",
     "karma-webpack": "^1.5.0",
-    "lodash": "^2.4.1",
+    "lodash": "^3.5.0",
     "merge-stream": "^0.1.7",
     "object-assign": "^2.0.0",
     "pretty-hrtime": "~0.2.1",
