+++ webpack.config.js
@@ -1,2 +1,2 @@
 // http://webpack.github.io/docs/list-of-plugins.html#commonschunkplugin
-module.exports = require('./gulpfile.js/config').webpack
+module.exports = require('./gulpfile.js/config/webpack')
